refactor: dropdown types
